Fixed problem submitting standard PAL 4:3 video (576i)
The height setter didn't include checking for 576 and thus any 4;3 video uploaded that was greater than 480 was upgraded to 720 and there isn't an entry in the 'workflow' DB, hence the failure.

There maybe issues with the heigh setter in future when screen-casts are added which may to support bespoke aspect ratios and sizes.

Signed-off-by: Ben Hawkridge <jamiedaniels123@gmail.com>