Fixed pagination problem regarding deleted tracks
Added an extra data selection condition to the pagination code.  In the case of a regular users (via index method) we only show tracks that haven't been deleted.

In the case of admin users using the special podcast view we use the method admin_index but this time allow the deleted flag to be 0 or 1.  A value of 2 means the track has been permanently deleted.

Signed-off-by: Ben Hawkridge <jamiedaniels123@gmail.com>