Track list view: fixed track image
Track image was failing to load because the permissionData method in the podcast model wasn't pulling out the 'custom_id' field and so the file path to the image wasn't valid.  Adding field fixes problem, re-enabled commented out view code to display images again.

Note, the code doesn't appear to default to the podcast image which it should do, need to review this and check areas like embedding tracks to ensure the behaviour is consistent.

Signed-off-by: Ben Hawkridge <jamiedaniels123@gmail.com>