Uploaded track media filenames should always be lowercase
This update makes sure that when the file is moved from the Filechucker upload folder to the admin transfer directory it converts the file to lowercase.  In the database at present the field 'Original_filename' will contain the original case and any sanitisation by Filechucker, meanwhile 'filename' contains the actual filename.

The code changes also move the logic of prefixing the filename to ensure 'uniqueness' to one location as the function 'moveFileChuckerUpload()' in folder.php also had the logic and it would have been easy to not realised it.

Signed-off-by: Ben Hawkridge <jamiedaniels123@gmail.com>