Player.xml bug fix
Restored the use of $this->podcast_item_media_folder in player.xml code but with bug fix.

On surface $this->podcast_item_media_folder is not needed for player.xml while only been generated for default flavour.  However in case other flavours are supported in future I have included $this->podcast_item_media_folder (as before) except this time taking into account that $this->podcast_item_media_folder has a leading '/' which means even for the default flavour it consists of a '/'.

This meant in the case of the "oup:longlink" atom link we have to remove the '/' as the shortcode is added to the url without a trailing '/'.

Signed-off-by: Ben Hawkridge <jamiedaniels123@gmail.com>