Added Duration column to the track list view
Added a duration column to the track list view, and a helper to format the duration (in seconds) to a modified SMPTE format.

Note, that we don't use PHP functions like date() as this take into account timezones and so will offset the time.  In theory gmdate() should work but current function is safer.

Signed-off-by: Ben Hawkridge <jamiedaniels123@gmail.com>