Multiple CakePHP script executions caused by files missing
(solved by disabled loading of type.css (doesn't exist) and added missing filechucker .js files)

In my continuing adventures with CakePHP I discovered that it was loading the pages three time (sort of).  Technically what was happening is that the default bootstrapping for CakePHP was firing off because it was loading something in the page (css, js, image etc) that it thought was actually a request for CakePHP rather than just a 'file'.

This issue isn't unique to CakePHP but applies to any framework that uses mod_rewrite rules to turn the URL into something you parser to then execute commands.  The basic formula in the apache config or .htaccess file is...

<IfModule mod_rewrite.c>
    RewriteEngine On
    RewriteCond %{REQUEST_FILENAME} !-d
    RewriteCond %{REQUEST_FILENAME} !-f
    RewriteRule ^(.*)$ index.php?url=$1 [QSA,L]
</IfModule>

...the two lines...

    RewriteCond %{REQUEST_FILENAME} !-d
    RewriteCond %{REQUEST_FILENAME} !-f

...essentially say, if there is NOT a real directory and there is NOT a real file then use the RewriteRule...

    RewriteRule ^(.*)$ index.php?url=$1 [QSA,L]

...which uses index.php in the web root and passes the current URL as a parameter.

The issue in our case was that two files...

type.css
/upload/js/js_table_sort.js

...didn't exist this then triggered the rewrite rules to pass the request via index.php, this then started up CakePHP executing various files until it eventually determined that it didn't know what to do with the request either and did nothing (404), meanwhile wasting resources.

There are some variations on the theme, for example this blog entry had problems with relative paths..

http://programming-perils.com/235/cakephp-script-keeps-executing-twice/
Signed-off-by: Ben Hawkridge <jamiedaniels123@gmail.com>