Session handling Changed - note core.php is now NOT ignored in GIT, this may affect deployment
As detailed on cakePHP web site at 'http://book.cakephp.org/1.3/en/view/1310/Sessions' I have added a custom implementation of the Session settings such that sessions are now destroyed when the browser closes.

Note that this required a change to 'app/config/core.php' which has previously been set to 'ignore' in GIT so this might have some impact from a deployment point of view.

For the dev version I've now set it to NOT be ignored.

Signed-off-by: Ben Hawkridge <jamiedaniels123@gmail.com>