(Soft) Deleting of tracks has now been fixed and some performance optimizations made
The (soft) delete method was not correctly (soft) deleting media files.  Soft deleting involves renaming the files so they are hidden (prefixed by period) and setting the 'delete' flag for the podcast_item track.

However the listAssociatedMedia() method was incorrectly assuming that each 'flavour' was in a sub-folder of that 'flavour' when the association is not that simply.  The actual association is defined in the $media_folder which was formerly in the feeds_controller but has now been moved the App_controller so it's available to all controllers.

In addition the updating of the RSS feeds has been moved to be outside the loop for (soft) deleting each track as it was generating far too many updates unnecessarily.  As a further optimisation if all the tracks were unpublished only the default flavour was refreshed as only player.xml actually needed updating.

NOTE: There are still a range of issues/todo's pending, however the primary concern is that poster images are not been (soft) deleted and there are possibly some issues with transcripts not been handled correctly.
Signed-off-by: Ben Hawkridge <jamiedaniels123@gmail.com>