CakePHP Sessions and SAMS (additional note about OUCU enforcement experiments)
Some changes to CakePHP Sessions such that they are now browser session based, this makes it easier to exit a session and helps ensure that the new session uses the current SAMS user, whereas before it would ignore it unless you actually destroyed the Session using the Podcast Logout.

Experimented with checked SAMS OUCU with Session OUCU however there are some issues when a user is in the server root, because the .htaccess mod_rewrite doesn't seem to correctly trigger the SAMS directive which only protects the index.php file, even though that file is loaded by Apache.  This failure at what might seem like random moments, would log the user out if we enforced the SAMS OUCU matching the Session OUCU

Signed-off-by: Ben Hawkridge <jamiedaniels123@gmail.com>