Some array_push()'s disabled as was causing PHP warnings
Some array_push commands were generating warnings due to NULL source arrays, however the lines appear to be redundant as the following line of code appears to do the same job.  Disabling doesn't appear to have broken anything, though there is an issue with the iTunes U category arrays not saving, though it's not thought to be related to this section of code.

Signed-off-by: Ben Hawkridge <jamiedaniels123@gmail.com>